Version 1.4.1

+++ system/info.go
@@ -305,7 +305,7 @@ func GetFSInfo() (map[string]*FSInfo, error) {
 		}
 
 		fsInfo.Total = stats.Blocks * uint64(stats.Bsize)
-		fsInfo.Free = stats.Bavail * uint64(stats.Bsize)
+		fsInfo.Free = uint64(stats.Bavail) * uint64(stats.Bsize)
 		fsInfo.Used = fsInfo.Total - (stats.Bfree * uint64(stats.Bsize))
 		fsInfo.IOStats = ios[strings.Replace(fsInfo.Device, "/dev/", "", 1)]
